Rename dashboard to update
