--- frontend/src/features/dashboard/routes/index.tsx
@@ -1,13 +1,0 @@
-import { Navigate, Route, Routes } from "react-router-dom";
-
-import { Dashboard } from "./Dashboard";
-
-export const DashboardRoutes = () => {
-  return (
-    <Routes>
-      <Route path="/" element={<Dashboard />} />
-    </Routes>
-  );
-};
-
-export { Dashboard } from "./Dashboard";
