+++ frontend/src/features/update/api/deleteUserSkill.ts
