+++ frontend/src/features/update/components/SkillCategoryKnown.tsx
