+++ frontend/src/routes/protected.tsx
@@ -2,7 +2,7 @@ import { Suspense } from "react";
 import { Outlet } from "react-router-dom";
 import { MainLayout } from "../components/Layout";
 import { Navigate } from "react-router-dom";
-import { DashboardRoutes } from "../features/dashboard";
+import { UpdateRoutes } from "../features/update";
 import { GraphRoutes } from "../features/graph";
 
 const App = () => {
@@ -28,7 +28,7 @@ export const protectedRoutes = [
     children: [
       {
         path: "/update",
-        element: <DashboardRoutes />,
+        element: <UpdateRoutes />,
       },
       {
         path: "/graph",
